# Performance optimizations for streaming top N rank