--- a/core/trino-main/src/main/java/io/trino/operator/GroupedTopNRankBuilder.java
+++ b/core/trino-main/src/main/java/io/trino/operator/GroupedTopNRankBuilder.java
@@ -91,7 +91,6 @@ public class GroupedTopNRankBuilder
                         return equalsAndHash.hashCode(page, position);
                     }
                 },
-                RowReferencePageManager.RESERVED_UNUSED_ROW_ID,
                 topN,
                 pageManager::dereference);
     }