--- a/core/trino-main/src/main/java/io/trino/operator/GroupedTopNRowNumberBuilder.java
+++ b/core/trino-main/src/main/java/io/trino/operator/GroupedTopNRowNumberBuilder.java
@@ -20,7 +20,6 @@ import io.trino.operator.RowReferencePageManager.LoadCursor;
 import io.trino.spi.Page;
 import io.trino.spi.PageBuilder;
 import io.trino.spi.type.Type;
-import it.unimi.dsi.fastutil.longs.LongComparator;
 import org.openjdk.jol.info.ClassLayout;
 
 import java.util.Iterator;
@@ -99,34 +98,15 @@ public class GroupedTopNRowNumberBuilder
     private void processPage(Page newPage, GroupByIdBlock groupIds)
     {
         try (LoadCursor loadCursor = pageManager.add(newPage)) {
-            GroupedTopNRowNumberAccumulator.RowReference rowReferenceView = asRowReferenceView(loadCursor);
             for (int position = 0; position < newPage.getPositionCount(); position++) {
                 long groupId = groupIds.getGroupId(position);
                 loadCursor.advance();
-                groupedTopNRowNumberAccumulator.add(groupId, rowReferenceView);
+                groupedTopNRowNumberAccumulator.add(groupId, loadCursor);
             }
             verify(!loadCursor.advance());
         }
 
         pageManager.compactIfNeeded();
-    }
-
-    private static GroupedTopNRowNumberAccumulator.RowReference asRowReferenceView(LoadCursor cursor)
-    {
-        return new GroupedTopNRowNumberAccumulator.RowReference()
-        {
-            @Override
-            public int compareTo(RowIdComparisonStrategy strategy, long rowId)
-            {
-                return cursor.compareTo(strategy, rowId);
-            }
-
-            @Override
-            public long extractRowId()
-            {
-                return cursor.allocateRowId();
-            }
-        };
     }
 
     private class ResultIterator