--- a/core/trino-main/src/main/java/io/trino/operator/RowIdComparisonHashStrategy.java
+++ b/core/trino-main/src/main/java/io/trino/operator/RowIdComparisonHashStrategy.java
@@ -1,0 +1,27 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.trino.operator;
+
+/**
+ * Strategy that combines both comparison and hash/equality over row IDs
+ */
+public interface RowIdComparisonHashStrategy
+        extends RowIdComparisonStrategy, RowIdHashStrategy
+{
+    @Override
+    default boolean equals(long leftRowId, long rightRowId)
+    {
+        return compare(leftRowId, rightRowId) == 0;
+    }
+}