… core/trino-main/src/main/java/io/trino/operator/TopNPeerGroupLookup.java
… core/trino-main/src/test/java/io/trino/operator/TestGroupedTopNRankAccumulator.java
… core/trino-main/src/test/java/io/trino/operator/TestGroupedTopNRowNumberAccumulator.java
… core/trino-main/src/test/java/io/trino/operator/TestRowReferencePageManager.java
… core/trino-main/src/test/java/io/trino/operator/TestTopNPeerGroupLookup.java