More complex performance improvements for all categories of top N rank usage, ranging up to over 70% speed up.

BEFORE:
Benchmark                             (groupCount)  (positions)  (topN)   Mode  Cnt  Score   Error  Units
BenchmarkGroupedTopNRankBuilder.topN             1      1000000       1  thrpt    7  7.442 ± 0.692  ops/s
BenchmarkGroupedTopNRankBuilder.topN             1      1000000      10  thrpt    7  7.508 ± 0.681  ops/s
BenchmarkGroupedTopNRankBuilder.topN             1      1000000     100  thrpt    7  6.999 ± 0.771  ops/s
BenchmarkGroupedTopNRankBuilder.topN         10000      1000000       1  thrpt    7  4.990 ± 0.720  ops/s
BenchmarkGroupedTopNRankBuilder.topN         10000      1000000      10  thrpt    7  1.389 ± 0.142  ops/s
BenchmarkGroupedTopNRankBuilder.topN         10000      1000000     100  thrpt    7  0.293 ± 0.015  ops/s
BenchmarkGroupedTopNRankBuilder.topN       1000000      1000000       1  thrpt    7  0.679 ± 0.042  ops/s
BenchmarkGroupedTopNRankBuilder.topN       1000000      1000000      10  thrpt    7  0.675 ± 0.025  ops/s
BenchmarkGroupedTopNRankBuilder.topN       1000000      1000000     100  thrpt    7  0.674 ± 0.014  ops/s

AFTER
Benchmark                             (groupCount)  (positions)  (topN)   Mode  Cnt  Score   Error  Units
BenchmarkGroupedTopNRankBuilder.topN             1      1000000       1  thrpt    8  9.491 ± 0.487  ops/s => +27.5% faster
BenchmarkGroupedTopNRankBuilder.topN             1      1000000      10  thrpt    8  9.422 ± 0.655  ops/s => +25.6% faster
BenchmarkGroupedTopNRankBuilder.topN             1      1000000     100  thrpt    8  9.159 ± 0.497  ops/s => +30.9% faster
BenchmarkGroupedTopNRankBuilder.topN         10000      1000000       1  thrpt    8  6.530 ± 0.561  ops/s => +30.9% faster
BenchmarkGroupedTopNRankBuilder.topN         10000      1000000      10  thrpt    8  1.568 ± 0.059  ops/s => +12.9% faster
BenchmarkGroupedTopNRankBuilder.topN         10000      1000000     100  thrpt    8  0.363 ± 0.004  ops/s => +23.9% faster
BenchmarkGroupedTopNRankBuilder.topN       1000000      1000000       1  thrpt    8  1.158 ± 0.051  ops/s => +70.5% faster
BenchmarkGroupedTopNRankBuilder.topN       1000000      1000000      10  thrpt    8  1.120 ± 0.115  ops/s => +65.9% faster
BenchmarkGroupedTopNRankBuilder.topN       1000000      1000000     100  thrpt    8  1.158 ± 0.153  ops/s => +71.8% faster