@sopel39 2021-01-18
[Benchmarks comparison-rank_v2.pdf](https://github.com/trinodb/trino/files/5829405/Benchmarks.comparison-rank_v2.pdf)

Not much of a difference in CPU